Refactor Markdown parsing out into a seperate file for reuse.
